requirements.txt
google-genai
pydantic>=2
ttkbootstrap
Pillow
